+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.12.1
+
+- Improved `sfreleaser changelog extract-section` to read for a github repository file directly.
+
+- Improved `sfreleaser changelog extract-section` to accept `--github-output=changelog:$GITHUB_OUTPUT`
+
 ## v0.12.0
 
 - Added `sfreleaser changelog extract-section` to be usable in GitHub CI for easy external release management.